+++ src/tsx/components/Header.tsx
@@ -1,16 +1,8 @@
 import React from "react";
-import logo from "../img/logo.svg";
+import logo from "../../img/logo.svg";
 import styled from "styled-components";
 
-const WEEKDAY = [
-  "일요일",
-  "월요일",
-  "화요일",
-  "수요일",
-  "목요일",
-  "금요일",
-  "토요일",
-];
+const WEEKDAY = ["일요일", "월요일", "화요일", "수요일", "목요일", "금요일", "토요일"];
 const INCREMENT = 1;
 const CHAR_COUNT = 2;
 
@@ -23,8 +15,10 @@ const renderCurrentDate = () => {
   const date = currentDate.getDate();
   const day = WEEKDAY[currentDate.getDay()];
 
-  return `${year}. ${month.toString().padStart(CHAR_COUNT, "0")}. ${date.toString().padStart(CHAR_COUNT, "0")}. ${day}`;
-}
+  return `${year}. ${month.toString().padStart(CHAR_COUNT, "0")}. ${date
+    .toString()
+    .padStart(CHAR_COUNT, "0")}. ${day}`;
+};
 
 function Header() {
   return (
