+++ src/tsx/components/RollingContainer.tsx
@@ -11,17 +11,17 @@ function RollingContainer() {
 }
 
 const Container = styled.div`
-margin-top: 40px;
-display: flex;
-justify-content: space-between;
-`
+  margin-top: 40px;
+  display: flex;
+  justify-content: space-between;
+`;
 
 const TextBox = styled.div`
-width: 32.85em;
+  width: 32.85em;
   height: 3.43em;
   border: 1px solid #d2dae0;
   background-color: #f5f7f9;
   overflow: hidden;
-`
+`;
 
 export default RollingContainer;
